requirements.txt
python-telegram-bot==20.0
Flask==2.3.2
gunicorn==23.0.0
 